Implemented -S simulate option and ping spacing.
The simulate option replaces actual probing with a non-response after
a uniformly-distributed random delay.  Ping spacing allows the client to
specify the minimum amount of time between the sending of a probe and the
return of the result to the client.  Hence, the client can space out
sequential probes by specifying a spacing value