tweaked the gperf generation code a little to make the error a little more obvious.
moved the logic into a seperate file (mper_keywords.gperf.sh) which means that
the warning messages are not dumped to the console unless they actually need to
be.

i'm still not entirely happy with this setup becuase the warning is dumped
inline with all the other stuff that is being spewed, so you'd have to be really
looking to notice it.